requirements-dev.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==7.1.2
Flask==1.1.4
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
pyee==7.0.4
python-dotenv==0.19.0
slackclient==2.9.3
slackeventsapi==2.2.1
typing-extensions==3.10.0.0
Werkzeug==1.0.1
yarl==1.6.3
requests==2.26.0
